+++ examples/json_serializable/lib/serializable.auto_mappr.dart
@@ -20,6 +20,7 @@ class $Mappr implements _i1.AutoMapprInterface {
   const $Mappr();
 
   Type _typeOf<T>() => T;
+
   List<_i1.AutoMapprInterface> get _delegates => const [];
 
   /// {@macro AutoMapprInterface:canConvert}
@@ -69,16 +70,23 @@ class $Mappr implements _i1.AutoMapprInterface {
   /// {@macro AutoMapprInterface:tryConvert}
   /// {@macro package:examples_json_serializable/serializable.dart}
   @override
-  TARGET? tryConvert<SOURCE, TARGET>(SOURCE? model) {
+  TARGET? tryConvert<SOURCE, TARGET>(
+    SOURCE? model, {
+    void Function(Object error, StackTrace stackTrace, SOURCE? source)?
+        onMappingError,
+  }) {
     if (canConvert<SOURCE, TARGET>(recursive: false)) {
-      return _convert(
+      return _safeConvert(
         model,
-        canReturnNull: true,
+        onMappingError: onMappingError,
       );
     }
     for (final mappr in _delegates) {
       if (mappr.canConvert<SOURCE, TARGET>()) {
-        return mappr.tryConvert(model);
+        return mappr.tryConvert(
+          model,
+          onMappingError: onMappingError,
+        );
       }
     }
 
@@ -108,13 +116,20 @@ class $Mappr implements _i1.AutoMapprInterface {
   /// {@macro package:examples_json_serializable/serializable.dart}
   @override
   Iterable<TARGET?> tryConvertIterable<SOURCE, TARGET>(
-      Iterable<SOURCE?> model) {
+    Iterable<SOURCE?> model, {
+    void Function(Object error, StackTrace stackTrace, SOURCE? source)?
+        onMappingError,
+  }) {
     if (canConvert<SOURCE, TARGET>(recursive: false)) {
-      return model.map<TARGET?>((item) => _convert(item, canReturnNull: true));
+      return model.map<TARGET?>(
+          (item) => _safeConvert(item, onMappingError: onMappingError));
     }
     for (final mappr in _delegates) {
       if (mappr.canConvert<SOURCE, TARGET>()) {
-        return mappr.tryConvertIterable(model);
+        return mappr.tryConvertIterable(
+          model,
+          onMappingError: onMappingError,
+        );
       }
     }
 
@@ -143,13 +158,23 @@ class $Mappr implements _i1.AutoMapprInterface {
   ///
   /// {@macro package:examples_json_serializable/serializable.dart}
   @override
-  List<TARGET?> tryConvertList<SOURCE, TARGET>(Iterable<SOURCE?> model) {
+  List<TARGET?> tryConvertList<SOURCE, TARGET>(
+    Iterable<SOURCE?> model, {
+    void Function(Object error, StackTrace stackTrace, SOURCE? source)?
+        onMappingError,
+  }) {
     if (canConvert<SOURCE, TARGET>(recursive: false)) {
-      return tryConvertIterable<SOURCE, TARGET>(model).toList();
+      return tryConvertIterable<SOURCE, TARGET>(
+        model,
+        onMappingError: onMappingError,
+      ).toList();
     }
     for (final mappr in _delegates) {
       if (mappr.canConvert<SOURCE, TARGET>()) {
-        return mappr.tryConvertList(model);
+        return mappr.tryConvertList(
+          model,
+          onMappingError: onMappingError,
+        );
       }
     }
 
@@ -178,13 +203,23 @@ class $Mappr implements _i1.AutoMapprInterface {
   ///
   /// {@macro package:examples_json_serializable/serializable.dart}
   @override
-  Set<TARGET?> tryConvertSet<SOURCE, TARGET>(Iterable<SOURCE?> model) {
+  Set<TARGET?> tryConvertSet<SOURCE, TARGET>(
+    Iterable<SOURCE?> model, {
+    void Function(Object error, StackTrace stackTrace, SOURCE? source)?
+        onMappingError,
+  }) {
     if (canConvert<SOURCE, TARGET>(recursive: false)) {
-      return tryConvertIterable<SOURCE, TARGET>(model).toSet();
+      return tryConvertIterable<SOURCE, TARGET>(
+        model,
+        onMappingError: onMappingError,
+      ).toSet();
     }
     for (final mappr in _delegates) {
       if (mappr.canConvert<SOURCE, TARGET>()) {
-        return mappr.tryConvertSet(model);
+        return mappr.tryConvertSet(
+          model,
+          onMappingError: onMappingError,
+        );
       }
     }
 
@@ -217,6 +252,35 @@ class $Mappr implements _i1.AutoMapprInterface {
           (model as _i2.ValueHolderDto?)) as TARGET);
     }
     throw Exception('No ${model.runtimeType} -> $targetTypeOf mapping.');
+  }
+
+  TARGET? _safeConvert<SOURCE, TARGET>(
+    SOURCE? model, {
+    void Function(Object error, StackTrace stackTrace, SOURCE? source)?
+        onMappingError,
+  }) {
+    if (!useSafeMapping<SOURCE, TARGET>()) {
+      return _convert(
+        model,
+        canReturnNull: true,
+      );
+    }
+    try {
+      return _convert(
+        model,
+        canReturnNull: true,
+      );
+    } catch (e, s) {
+      onMappingError?.call(e, s, model);
+      return null;
+    }
+  }
+
+  /// {@macro AutoMapprInterface:useSafeMapping}
+  /// {@macro package:examples_json_serializable/serializable.dart}
+  @override
+  bool useSafeMapping<SOURCE, TARGET>() {
+    return false;
   }
 
   _i2.User _map__i2$UserDto_To__i2$User(_i2.UserDto? input) {